refactored students list
