+++ client/src/components/home/ta/TAMain.js
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import AskQuestion from './AskQuestion';
-import StudentEntries from './StudentEntries';
+import StudentEntries from '../shared/StudentEntries';
 
 function TAMain (props) {
     const { theme, queueData } = props;
